fix: Keep vector distances in float64 to avoid overflow
Squaring a large float32 component exceeds what a float32 holds, so the
distance saturated to +Inf. Two such documents compared equal, leaving
their order arbitrary.

Resolves #5220

+++ internal/index/hnsw/search.go
@@ -44,7 +44,7 @@ func (g *HNSWIndex) SearchWithDistance(query []float32, k, efSearch int) ([]Neig
 	}
 	out := make([]Neighbor, len(w))
 	for i, c := range w {
-		out[i] = Neighbor{ID: c.id, Distance: float64(c.dist)}
+		out[i] = Neighbor{ID: c.id, Distance: c.dist}
 	}
 	return out, nil
 }
